Make sure module can be used as main programm
Python has this interesting idiom "if __name__ == __main__", which
determines what to do when a module is imported or directly called.

It is odd to test, though, as - when importing the module into
your test suite - you cannot reach the other branch.

There are different workarounds available. Thanks to Steve for his
answer about how to test this idiom:
https://mail.python.org/pipermail/tutor/2016-February/108119.html

With this test, the coverage now has reached 100%!!

Still, tests have to be improved, as some call the live API and
other issues.

modified:   README.md
new file:   tests/test_main_programm.py